0521-1039 修改配置文件中加密信息的存储方式（字符改16进制数）

+++ PCBDetect/Configurator.h
@@ -193,7 +193,7 @@ namespace pcb
 	private:
 		QFile *configFile;
 		QDateTime fileDateTime;
-		int keys[4]; //√ÿ‘ø
+		ushort keys[4]; //√ÿ‘ø
 
 	public:
 		Configurator(QFile *file = Q_NULLPTR);
